refactor: make it work on dep again
