+++ go.mod
@@ -1,3 +1,6 @@
 module github.com/caarlos0/env/v5
 
-require github.com/stretchr/testify v1.3.0
+require (
+	github.com/caarlos0/env v3.5.0+incompatible
+	github.com/stretchr/testify v1.3.0
+)
